third_party/WebKit/LayoutTests/media/track/track-remove-track.html

<!DOCTYPE html>
<html>
    <head>
        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />

        <script src=../media-file.js></script>
        <script src=../video-test.js></script>
        <script>

            var trackElement;
            var track;

            function trackRemoved()
            {
                consoleWrite("EVENT(" + event.type + ")");

                testExpected("event.target", video.textTracks);
                testExpected("event instanceof window.TrackEvent", true);
                testExpected("event.track", track);

                consoleWrite("");
                endTest();
            }

            function start()
            {
                findMediaElement();

                consoleWrite("** Test a TextTrack created for a &lt;track&gt; element.");
                run("trackElement = document.createElement('track')");
                run("video.appendChild(trackElement)");

                run("trackElement.src = 'captions-webvtt/tc004-webvtt-file.vtt'");
                run("trackElement.track.mode = 'hidden'");
                testExpected("video.textTracks.length", 1);

                track = video.textTracks[0];
                video.removeChild(trackElement);

                video.textTracks.addEventListener("removetrack", trackRemoved);
            }

        </script>
    </head>
    <body onload="start()">
        <p>Tests that the 'removetrack' event is fired when a TextTrack is removed.</p>
        <video>
        </video>
    </body>
</html>
