third_party/WebKit/LayoutTests/media/track/track-remove-track-expected.txt

Tests that the 'removetrack' event is fired when a TextTrack is removed.

** Test a TextTrack created for a <track> element.
RUN(trackElement = document.createElement('track'))
RUN(video.appendChild(trackElement))
RUN(trackElement.src = 'captions-webvtt/tc004-webvtt-file.vtt')
RUN(trackElement.track.mode = 'hidden')
EXPECTED (video.textTracks.length == '1') OK
EVENT(removetrack)
EXPECTED (event.target == '[object TextTrackList]') OK
EXPECTED (event instanceof window.TrackEvent == 'true') OK
EXPECTED (event.track == '[object TextTrack]') OK

END OF TEST
